# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Este es el proyecto final de la materia espcialización backend 3 para crear, editar, consultar y borrar las entidades de paciente, dentista y turnos.",
        "title": "Proyecto Final Especialización backend 3",
        "contact": {
            "name": "API Support (Natalia Garcia, Federico Bustamante, Damian, Camilo Zuleta)"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http:www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/dentists": {
            "post": {
                "description": "Creates a Dentist, you must send the fields required to process your request they are name, surname and license",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dentist"
                ],
                "summary": "Creates a Dentist",
                "parameters": [
                    {
                        "type": "string",
                        "description": "publicKey",
                        "name": "PUBLIC-KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "secretKey",
                        "name": "SECRET_KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "Create Dentist",
                        "name": "Dentist",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dentists.Dentist"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dentists.Dentist"
                        }
                    }
                }
            }
        },
        "/dentists/{id}": {
            "get": {
                "description": "Gets a Dentist by id using the repository principal",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dentist"
                ],
                "summary": "Gets a Dentist by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "publicKey",
                        "name": "PUBLIC-KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "secretKey",
                        "name": "SECRET_KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dentists.Dentist"
                        }
                    }
                }
            },
            "put": {
                "description": "Updates a Dentist by id, you must send all of the dentist fields to process your request",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dentist"
                ],
                "summary": "Updates a Dentist by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "publicKey",
                        "name": "PUBLIC-KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "secretKey",
                        "name": "SECRET_KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "update Dentist",
                        "name": "Dentist",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dentists.Dentist"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dentists.Dentist"
                        }
                    }
                }
            },
            "delete": {
                "description": "Deletes a Dentist by ID, be careful with this option!",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dentist"
                ],
                "summary": "Deletes a Dentist by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "publicKey",
                        "name": "PUBLIC-KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "secretKey",
                        "name": "SECRET_KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dentists.Dentist"
                        }
                    }
                }
            },
            "patch": {
                "description": "Updates a Dentist by id, you must send all of the dentist required fields to process your request",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Dentist"
                ],
                "summary": "Updates a Dentist by id",
                "parameters": [
                    {
                        "type": "string",
                        "description": "publicKey",
                        "name": "PUBLIC-KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "secretKey",
                        "name": "SECRET_KEY",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "update Dentist",
                        "name": "Dentist",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dentists.Dentist"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dentists.Dentist"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dentists.Dentist": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer"
                },
                "license": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "surname": {
                    "type": "string"
                }
            }
        },
...
    }
}